Updated the design, started work on the search
Still need to put in the links for the nav at the top, fix the weird location show bug, and hook up the rest of the search. There's a function to do the search in map.js.coffee, it just needs to be finished so that it actually does the ajax request to do the search, and then hides all the locations that are not returned ( or hide everything, and show what came back in the search, either way works. )